+++ CHANGELOG.md
@@ -7,7 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
-## [0.45.0] - 2024-07-29
+## [0.45.0] - 2024-07-31
+
+### Breaking changes
+
+-   Updated the font and font-weights in the default styles
+-   We no longer load the Rubik font by default
 
 ### Fixes